+++ src/Database.php
@@ -33,6 +33,17 @@ class Database
         'charset' => 'utf8',
     ];
     private static $opt;
+    private static $count = 0;
+
+    /**
+     * @var bool Default value for query cache status
+     */
+    private static $cache = FALSE;
+
+    /**
+     * @var array Storage var for query cache data
+     */
+    public static $cacheStorage = [];
 
     /**
      * Validate database connection type
@@ -238,6 +249,8 @@ class Database
             return FALSE;
         }
 
+        self::$count = 0;
+
         return TRUE;
     }
 
@@ -262,12 +275,14 @@ class Database
         // get \PDOStatement
         $stmt = Connection::SQL($sql, $params);
 
+        self::$count++;
+
         $sqlSubStr6 = mb_substr($sql, 0, 6);
 
         if ($sqlSubStr6 == 'SELECT') {
             // return fetch array
             return $stmt->fetchAll(\PDO::FETCH_ASSOC);
-        } elseif ($sqlSubStr6 == 'INSERT' || $sqlSubStr6 == 'UPDATE') {
+        } elseif ($sqlSubStr6 == 'INSERT' || $sqlSubStr6 == 'UPDATE' || $sqlSubStr6 == 'DELETE') {
             // return row count
             return $stmt->rowCount();
         } else {
@@ -277,9 +292,75 @@ class Database
         return $result;
     }
 
+    /**
+     * @since  0.4
+     *
+     * @return int
+     */
+    public static function getCount()
+    {
+        return self::$count;
+    }
+
     public static function Query()
     {
         return new QueryMain();
+    }
+
+    /**
+     * @since  0.4
+     * @return bool
+     */
+    public static function getCacheDefault()
+    {
+        return self::$cache;
+    }
+
+    /**
+     * Set default cache status
+     *
+     * @param bool $cache
+     *
+     * @since  0.4
+     * @return bool
+     * @throws \Exception
+     */
+    public static function setCacheDefault($cache)
+    {
+        if (!is_bool($cache)) {
+            // cache status must be a bool value
+            throw new \Exception('fly\Database: Expects parameter 1 to be a valid default cache status');
+        }
+
+        // set default cache status
+        self::$cache = $cache;
+
+        // return actual status
+        return self::getCacheDefault();
+    }
+
+    /**
+     * Set default cache status to TRUE
+     *
+     * @since  0.4
+     * @return bool
+     * @throws \Exception
+     */
+    public static function setCacheDefaultTrue()
+    {
+        return self::setCacheDefault(TRUE);
+    }
+
+    /**
+     * Set default cache status to FALSE
+     *
+     * @since  0.4
+     * @return bool
+     * @throws \Exception
+     */
+    public static function setCacheDefaultFalse()
+    {
+        return self::setCacheDefault(FALSE);
     }
 
 }