+++ src/database/QueryBuilder.php
@@ -204,50 +204,56 @@ class QueryBuilder extends QueryValidator
 
     /**
      * Add field or fields for select query
-     * Format: '*' OR 'field' OR ['field1', 'field2', ... ]
      *
-     * @param string|array $data
+     * Format: 'field'
+     * Format: ['field1', 'field2', ... ]
+     * Format: [ ['func()', 'alias'], ['table.field', 'alias'], ['field', 'alias'], ['table.field'], ['field'] ]
+     *
+     * @param string|array $input
      *
      * @return $this
      * @throws \Exception
      */
-    public function addSelect($data = '*')
+    public function addSelect($input = '*')
     {
         // Check query type
         if (!$this->_checkQueryTypeAvailableAndSetSelect()) {
             throw new \Exception('fly\Database: Query type is not a SELECT');
         }
 
-        // TODO: rewrite this part of method
+        if (!ArrayHelper::isArray($input) && is_string($input)) {
+            // Format: 'field'
+            // Reformat to 2D strong array
+            $input = [[trim($input)]];
+        }
 
-        if (ArrayHelper::isArrayStrong($data)) {
-            // it is one-level array of new fields
-            foreach ($data as $field) {
-                if ($this->_isFieldNameValid($field)) {
-                    $this->select[] = $field;
-                } else {
-                    throw new \Exception('fly\Database: Field name in method addSelect() is not valid');
+        if (ArrayHelper::isArrayStrong($input)) {
+            // Format: ['field1', 'field2', ... ]
+            // Reformat to 2D strong array
+            foreach ($input as $key => $inputItem) {
+                $input[$key] = [trim($inputItem)];
+            }
+        }
+
+        if (ArrayHelper::isArray2DStrong($input)) {
+            // Format: [ ['func()', 'alias'], ['table.field', 'alias'], ['field', 'alias'], ['table.field'], ['field'] ]
+            foreach ($input as $inputItem) {
+                if (ArrayHelper::isArray($inputItem)) {
+                    if (count($inputItem) === 2) {
+                        $field = array_shift($inputItem);
+                        $alias = array_shift($inputItem);
+                        $this->select[] = [$field, $alias];
+                    } elseif (count($inputItem) === 1) {
+                        $field = array_shift($inputItem);
+                        $this->select[] = [$field];
+                    } else {
+                        throw new \Exception('fly\Database: Invalid input data for SELECT part');
+                    }
                 }
             }
-
-            return $this;
         }
 
-        if (!ArrayHelper::isArray($data) && $this->_isFieldNameValid($data)) {
-            // it is a one field
-            $this->select[] = $data;
-
-            return $this;
-        }
-
-        if ($data === '*') {
-            // all fields
-            $this->select = ['*'];
-
-            return $this;
-        }
-
-        throw new \Exception('fly\Database: Expects parameter 1 to be a valid data array');
+        return $this;
     }
 
     /**
@@ -392,7 +398,7 @@ class QueryBuilder extends QueryValidator
                 $row[] = '=';
             }
 
-            if (count($row) === 3 && in_array($row[2], ['=', '>', '<', '>=', '<=', '<>'])) {
+            if (count($row) === 3 && in_array($row[2], ['=', '>', '<', '>=', '<=', '<>', 'LIKE', 'NOT LIKE'])) {
                 // format: ['field', 'value', '?']
                 list($fieldName, $fieldValue, $fieldOperator) = $row;
                 if ($this->_isFieldNameValid($fieldName)) {
@@ -407,14 +413,17 @@ class QueryBuilder extends QueryValidator
 
         if (ArrayHelper::isArray2D($row)) {
             // two-level array
-            if (count($row) === 2 && ArrayHelper::isArray($row[1]) && count($row[1])) {
+            if ((count($row) === 2 || count($row) === 3) && ArrayHelper::isArray($row[1]) && count($row[1]) && (!isset($row[2]) || in_array($row[2], ['IN', 'NOT IN']))) {
                 // format: ['field', ['value1', 'value2', 'value3']]
+                // format: ['field', ['value1', 'value2', 'value3'], 'IN']
+                // format: ['field', ['value1', 'value2', 'value3'], 'NOT IN']
                 list($fieldName, $fieldValues) = $row;
+                $fieldOperator = $row[2] ?? 'IN';
                 if ($this->_isFieldNameValid($fieldName)) {
                     return [
                         $fieldName,
                         $fieldValues,
-                        'IN',
+                        $fieldOperator,
                     ];
                 }
             }