+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+## v0.4 (June 27, 2018)
+- Add SQL counter to `fly\Database`.
+- Add query cache to `fly\Database`.
+- Add function support in select part to Query Builder.
+- Add 'NOT IN', 'LIKE', 'NOT LIKE' support to Query Builder.
+- Add multi file support for config in class `fly\Route`.
+- Add custom 'Error 403. Forbidden' page in class `fly\Route`.
+- Add custom 'Error 404. Object not found' page in class `fly\Route`.
+
 ## v0.3 (March 15, 2018)
 - Add 'allow' option for config in class `fly\Route`.
 - Add 'Error 403. Forbidden' page in class `fly\Route`.